No longer responds to bots

+++ bot.py
@@ -494,24 +494,26 @@ def on_message(message):
         print(time.strftime("%Y-%m-%d %H:%M:%S") + ": " + message.author.name + " sent something containing emojis.")
     if message.author == client.user:
         return
-    #Uncomment this if you want to use an exclamaion mark instead of @pybot
-    if message.content.startswith("!") and len(message.content) > 1:
-        yield from processCommand(message.content[1:], message.channel, message.author, message)
-    if message.content.startswith("<@" + client.user.id + ">") and len(message.content) > 22:
-        yield from client.send_typing(message.channel)
-        yield from processCommand(str.strip(message.content[22:]), message.channel, message.author, message)
-    if message.content.startswith("@" + client.user.name) and len(message.content) > 7:
-        yield from client.send_typing(message.channel)
-        yield from processCommand(str.strip(message.content[7:]), message.channel, message.author, message)
-    elif message.content.startswith("-p"):
-        yield from client.send_typing(message.channel)
-        yield from processCommand(str.strip(message.content[3:]), message.channel, message.author, message)
-    elif message.content == "<@" + client.user.id + ">" or message.content == "-p" or message.content == "@" + client.user.name:
-        yield from client.send_message(message.channel, 'Hello {}!'.format(message.author.mention))
-    elif message.content == "sudo rm -rf":
-        yield from processCommand("clearideas", message.channel, message.author, message)
-    else:
-        yield from processResponse(message)
+    if "NadekoFlower" in message.content:
+        yield from client.send_message(message.channel, ">pick")    
+    if message.author.bot == False:
+        if message.content.startswith("!") and len(message.content) > 1:
+            yield from processCommand(message.content[1:], message.channel, message.author, message)
+        if message.content.startswith("<@" + client.user.id + ">") and len(message.content) > 22:
+            yield from client.send_typing(message.channel)
+            yield from processCommand(str.strip(message.content[22:]), message.channel, message.author, message)
+        if message.content.startswith("@" + client.user.name) and len(message.content) > 7:
+            yield from client.send_typing(message.channel)
+            yield from processCommand(str.strip(message.content[7:]), message.channel, message.author, message)
+        elif message.content.startswith("-p"):
+            yield from client.send_typing(message.channel)
+            yield from processCommand(str.strip(message.content[3:]), message.channel, message.author, message)
+        elif message.content == "<@" + client.user.id + ">" or message.content == "-p" or message.content == "@" + client.user.name:
+            yield from client.send_message(message.channel, 'Hello {}!'.format(message.author.mention))
+        elif message.content == "sudo rm -rf":
+            yield from processCommand("clearideas", message.channel, message.author, message)
+        else:
+            yield from processResponse(message)
 @client.event
 @asyncio.coroutine
 def on_ready():
